Menú funcionando y generador tmb

+++ generador.py
@@ -8,7 +8,7 @@ def generar_imagen(n, m, salida_path):
     imagen = np.full((m, n, 3), fondo_color, dtype=np.uint8)
 
     # Calcular el tamaño máximo de las figuras geométricas
-    tamano_figura = min(n, m)/5
+    tamano_figura = int(min(n, m) / 5)  # Convertir a entero
 
     # Generar una secuencia aleatoria de figuras geométricas
     for _ in range(10):  # Número de figuras a dibujar
@@ -20,8 +20,8 @@ def generar_imagen(n, m, salida_path):
             cv2.circle(imagen, centro, radio, color_figura, -1)
         elif figura == 'rectangulo':
             esquina1 = (random.randint(0, n-1), random.randint(0, m-1))
-            esquina2 = (random.randint(esquina1[0], min(esquina1[0] + int(tamano_figura), n-1)),
-                        random.randint(esquina1[1], min(esquina1[1] + int(tamano_figura), m-1)))
+            esquina2 = (random.randint(esquina1[0], min(esquina1[0] + tamano_figura, n-1)),
+                        random.randint(esquina1[1], min(esquina1[1] + tamano_figura, m-1)))
             cv2.rectangle(imagen, esquina1, esquina2, color_figura, -1)
         elif figura == 'triangulo':
             p1 = (random.randint(0, n-1), random.randint(0, m-1))
